Move non-view classes to separate files
